feat : Member 객체 생성
