+++ src/main/java/jpa/entity/Member.java
@@ -1,0 +1,30 @@
+package jpa.entity;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.*;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@EntityListeners(AuditingEntityListener.class)
+public class Member extends BaseTimeEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    private long age;
+    private String email;
+    private String password;
+
+    public Member(long age, String email, String password) {
+        this.age = age;
+        this.email = email;
+        this.password = password;
+    }
+
+    public long getAge() {
+        return age;
+    }
+}
