+++ src/main/java/jpa/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package jpa.repository;
 
-import jpa.domain.Member;
+import jpa.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
